refactor(create-tour): separate component creatourstep2 to task1 and task2
