+++ 3d-virtual-nlu/src/components/admin/RightMenuCT.tsx
@@ -1,39 +1,52 @@
+import { FaLock } from "react-icons/fa6";
 import styles from "../../styles/rightmenu.module.css";
+import { MdDoneOutline } from "react-icons/md";
 /**
  * - Nhận thấy rằng step 2 & step 3 chia sẻ cùng UI.
  */
-type RightMenuProps = {
-  tasks: { id: number; title: string; content: React.ReactNode }[];
-  activeTaskId: number;
-  completedTaskIds: number[];
-  onSelectTask: (id: number) => void;
+
+interface TaskItem {
+  id: number;
+  title: string;
+}
+
+interface RightMenuProps {
+  // tasks: { id: number; title: string; content: React.ReactNode }[];
+  tasks: TaskItem[];
   openTaskIndex: number | null;
-  setOpenTaskIndex: (id: number) => void;
-};
+  unlockedTaskIds: number[];
+  completedTaskIds: number[];
+  onTaskClick: (id: number) => void;
+}
 
 const RightMenuCreateTour: React.FC<RightMenuProps> = ({
   tasks,
-  activeTaskId,
-  completedTaskIds,
   openTaskIndex,
-  setOpenTaskIndex,
-  onSelectTask,
+  unlockedTaskIds,
+  completedTaskIds,
+  onTaskClick,
 }) => {
   return (
     <>
       <ul>
         {tasks.map((task) => {
-          const isOpen = openTaskIndex === task.id;
-          const isUnlocked =
-            task.id === 1 || completedTaskIds.includes(task.id - 1);
+          const isUnlocked = unlockedTaskIds.includes(task.id);
+          const isCompleted = completedTaskIds.includes(task.id);
+          const isActive = openTaskIndex === task.id;
 
           return (
             <li
               key={task.id}
               className={styles.taskRow}
-              onClick={() => isUnlocked && onSelectTask(task.id)}
+              style={{
+                cursor: isUnlocked ? "pointer" : "not-allowed",
+                borderBottom: isActive ? "2px solid #7FFF00" : "",
+              }}
+              onClick={() => isUnlocked && onTaskClick(task.id)}
             >
               <span className={styles.taskName}>{task.title}</span>
+              {!isUnlocked && <FaLock className={styles.taskIcon} />}
+              {isCompleted && <MdDoneOutline className={styles.taskIcon} />}
             </li>
           );
         })}
