+++ 3d-virtual-nlu/src/components/admin/TaskContainerCT.tsx
@@ -29,6 +29,7 @@ const TaskContainerCT: React.FC<TaskContainerCTProps> = ({
         {id} : {name}
       </h3>
       <div className={styles.taskContent}>{children}</div>
+
       <button className={styles.taskSave} onClick={onSave}>
         <IoIosSave />
       </button>
